Changelog in description
execution_flow, hook_handler, print_override, library_loader, runtime_handler, variant_watcher uses internal storage for storing its own data.
Runtime objects constructor and destructor have their own export function for library.
Runtime objects constructor and destructor added to API abstraction object.
func_db removed from runtime_handler.
luastr_util is deprecated, therefore, the source code is deleted.
Modified README.txt in Src folder.
lualib_io_handler renamed to lualibrary_iohandler to accomodate the naming convention.
stdlogger renamed to std_logger to accomodate the naming convention.
Created functional copy of IO library of Lua counterpart

Fixed potential bug where library_loader might have incorrect algorithm for replacing current library object.
Fixed potential bug where object_var can push the object into another different state when it already exists in another state.
