+++ Src/luaapi_compilation_context.cpp
@@ -4,6 +4,7 @@
 #include "luaapi_internal.h"
 #include "luaapi_memory.h"
 #include "luaapi_object_util.h"
+#include "luaapi_runtime.h"
 #include "luaapi_stack.h"
 #include "luaapi_state.h"
 #include "luaapi_util.h"
@@ -17,12 +18,13 @@ using namespace lua::api;
 
 #ifdef LUA_CODE_EXISTS
 
-compilation_context::compilation_context(I_debug* debug, I_execution* execution, I_internal* internal, I_memory* memory, I_object_util* objutil, I_stack* stack, I_state* state, I_util* util, I_value* value, I_table_util* tableutil, I_variant_util* varutil){
+compilation_context::compilation_context(I_debug* debug, I_execution* execution, I_internal* internal, I_memory* memory, I_object_util* objutil, I_runtime* runtime, I_stack* stack, I_state* state, I_util* util, I_value* value, I_table_util* tableutil, I_variant_util* varutil){
   api_debug = debug;
   api_execution = execution;
   api_internal = internal;
   api_memory = memory;
   api_objutil = objutil;
+  api_runtime = runtime;
   api_stack = stack;
   api_state = state;
   api_util = util;
@@ -33,17 +35,18 @@ compilation_context::compilation_context(I_debug* debug, I_execution* execution,
 
 
 static const lua::api::compilation_context __compilation_context(
-  cpplua_get_api_debug_definition(),
-  cpplua_get_api_execution_defintion(),
-  cpplua_get_api_internal_definition(),
-  cpplua_get_api_memory_definition(),
-  cpplua_get_api_object_util_definition(),
-  cpplua_get_api_stack_definition(),
-  cpplua_get_api_state_definition(),
-  cpplua_get_api_util_definition(),
-  cpplua_get_api_value_definition(),
-  cpplua_get_api_table_util_definition(),
-  cpplua_get_api_variant_util_definition()
+  CPPLUA_GET_API_DEBUG_DEFINITION(),
+  CPPLUA_GET_API_EXECUTION_DEFINITION(),
+  CPPLUA_GET_API_INTERNAL_DEFINITION(),
+  CPPLUA_GET_API_MEMORY_DEFINITION(),
+  CPPLUA_GET_API_OBJECT_UTIL_DEFINITION(),
+  CPPLUA_GET_API_RUNTIME_DEFINITION(),
+  CPPLUA_GET_API_STACK_DEFINITION(),
+  CPPLUA_GET_API_STATE_DEFINITION(),
+  CPPLUA_GET_API_UTIL_DEFINITION(),
+  CPPLUA_GET_API_VALUE_DEFINITION(),
+  CPPLUA_GET_API_TABLE_UTIL_DEFINITION(),
+  CPPLUA_GET_API_VARIANT_UTIL_DEFINITION()
 );
 
 DLLEXPORT const lua::api::compilation_context* CPPLUA_GET_API_COMPILATION_CONTEXT(){
